VST-61377: Use Correct Validation Endpoint
Why?
The current validation method uses the old Universal Analytics Endpoint. We want be able to use the old and the new measurement protocol endpoint. This is purely for developer testing and has not impact on the current collection endpoint.

What?
Add an optional param to pass in the new measurement protocol validation endpoint.